requirements.txt
groq==0.33.0
ollama==0.6.1
python-dotenv==1.2.1
pydantic==2.12.5
pdf2image==1.17.0
pytesseract==0.3.13
requests==2.32.5